+++ packages/core/src/migrate/migrator.ts
@@ -65,7 +65,7 @@ export class Migrator {
         tableName: options.tableName,
       });
     this.migrationsPath = options.migrationsPath ?? './migrations';
-    this._logger = new LoggerWrapper(options.logger, options.slowQueryThreshold);
+    this._logger = new LoggerWrapper(options.logger, options.slowQuery);
     this._entities = options.entities;
     this.schemaIntrospector = this.createIntrospector();
     this.schemaGenerator = options.schemaGenerator ?? this.createGenerator(options.namingStrategy);